requirements.txt
fastapi
uvicorn[standard]
PyMuPDF
openai
transformers
torch
psycopg2-binary
pgvector
pydantic
pydantic-settings
python-dotenv
python-multipart
SQLModel
numpy
asyncio
logging